requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
GDAL==3.8.4
numpy==2.2.6
psycopg2==2.9.10
scipy==1.15.3
sqlparse==0.5.3
tqdm==4.67.1
